\name{celegansbio}
\docType{data}
\alias{celegansbio}
\title{
Bioluminescence of C. Elegans
}
\description{
 This dataset gives the time series of bioluminescence of an experiment monitoring C. Elegans as they feed and forage.  The observations are taken 6-minutes apart with no missing data.
}
\usage{
celegansbio
}
\format{
	A vector of length 623.
}
\source{Experiment from Alexandre Benedetto's research group at Lancaster University.}

\keyword{datasets}
