\name{celegensbio}
\docType{data}
\alias{celegensbio}
\title{
Bioluminescence of C. Elegens
}
\description{
 This dataset gives the time series of bioluminescence of an experiment monitoring C. Elegens as they feed and forage.  The observations are taken 6-minutes apart with no missing data.
}
\usage{
celegensbio
}
\format{
	A vector of length 623.
}
\source{Experiment from Alexandre Benedetto's research group at Lancaster University.}

\keyword{datasets}
