% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.boundary.timeseries.R
\name{get.boundary.timeseries}
\alias{get.boundary.timeseries}
\title{Calculate Boundary Extended Time Series}
\usage{
get.boundary.timeseries(x, type = c("TLSW", "LSW.diff")[1])
}
\arguments{
\item{x}{The time series used to calculate the boundary extended version.}

\item{type}{The type of boundary handling used. Either "TLSW" for boundary
handling as described in McGonigle, E. T., Killick, R., and Nunes, M. (2022)
Trend locally stationary wavelet processes, or "LSW.diff" for a periodic
version of this, used for the differencing-based functions.}
}
\value{
A vector of 4 times the length of the input vector.
}
\description{
A function to calculate the boundary extended time series, to be used within
the \code{\link{TLSW}} function.
Not recommended for general usage.
}
\examples{
trend <- seq(from = 0, to = 2, length = 400)
set.seed(1)

x <- rnorm(400) + trend

x.b <- get.boundary.timeseries(x)

plot.ts(x.b)
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend
locally stationary wavelet processes. \emph{Journal of Time Series
Analysis}, 43(6), 895-917.
}
\seealso{
\code{\link{TLSW}}
}
