% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacf.calc.R
\name{lacf.calc}
\alias{lacf.calc}
\title{Compute Localised Autocovariance Estimate from Spectrum Estimate}
\usage{
lacf.calc(
  x,
  filter.number = 4,
  family = "DaubExPhase",
  spec.est = NULL,
  lag.max = NULL,
  ...
)
}
\arguments{
\item{x}{The time series you wish to analyse.}

\item{filter.number}{Wavelet filter number that generated the time series.}

\item{family}{Wavelet family that generated the time series.}

\item{spec.est}{Estimated spectrum from which the lacf estimate will be
calculated, the component \code{spec.est} in the output of the \code{TLSW} functions.}

\item{lag.max}{The maximum lag of acf required. If NULL then the same
default as in the regular acf function is used.}

\item{...}{Further arguments to be passed to perform spectrum estimation, only to be used if \code{spec.est} is not supplied.}
}
\value{
An object of class \code{lacf} which contains the autocovariance.
}
\description{
Computes the local autocovariance and autocorrelation estimates, given an
input of a spectrum estimate. Provides the same functionality as the
function \code{lacf} from the \code{locits} package, but user provides the spectrum
estimate in the argument.
}
\examples{

## ---- computes estimate of local autocovariance function

## ---- example where LSW process is generated using the Haar wavelet

spec <- wavethresh::cns(512)
spec <- wavethresh::putD(spec, level = 8, 1 + sin(seq(from = 0, to = 2 * pi, length = 512))^2)

noise <- wavethresh::LSWsim(spec)
trend <- seq(from = 0, to = 10, length = 512)

x <- trend + noise

## ---- first estimate the spectrum using Daubechies EP4 wavelet:

x.TLSW <- TLSW(x)

spec.est <- x.TLSW$spec.est

#---- estimate the lacf:

lacf.est <- lacf.calc(x = x, spec.est = spec.est)

plot.ts(lacf.est$lacf[, 1])
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend
locally stationary wavelet processes. \emph{Journal of Time Series
Analysis}, 43(6), 895-917.

Nason, G. P. (2013). A test for second-order stationarity and approximate
confidence intervals for localized autocovariances for locally stationary
time series. \emph{Journal of the Royal Statistical Society: Series B
(Statistical Methodology)}, \bold{75(5)}, 879--904.
}
\seealso{
\code{\link{lacf}}
}
