% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSWsim.anydist.R
\name{LSWsim.anydist}
\alias{LSWsim.anydist}
\title{Simulate locally stationary wavelet processes with random innovations not
necessarily Gaussian.}
\usage{
LSWsim.anydist(spec, distribution = "Normal")
}
\arguments{
\item{spec}{An object of class wd which contains the spectrum for simulating
an LSW process.}

\item{distribution}{The distribution of the random variables used to
simulate the process. Can be "Normal", "Exponential", "Chisquare" or
"Poisson".}
}
\value{
A vector simulated from the spectral description given in the spec
description. The returned vector will exhibit the spectral characteristics
defined by spec.
\code{\link{LSWsim}}
}
\description{
Simulates a locally stationary wavelet process given a spectrum and
distribution for the random innovations. Extension of the LSWsim function
from wavethresh.
}
\examples{
spec <- wavethresh::cns(1024)

spec <- wavethresh::putD(spec, level = 8, seq(from = 2, to = 8, length = 1024))

x <- LSWsim.anydist(spec, distribution = "Exponential")

plot.ts(x)
}
