% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TLSW.R
\name{plot.TLSW}
\alias{plot.TLSW}
\title{Title}
\usage{
\method{plot}{TLSW}(x, plot.type = c("both", "trend", "spec")[1], ...)
}
\arguments{
\item{x}{A \code{TLSW} object}

\item{plot.type}{A string object indicating what is to be plotted. Can be "trend", in which case
the trend estimate (and associated confidence intervals if calculated) are plotted, or "spec",
in which case the spectral esitmate is plotted, or "both", in which case both estimates are plotted.}

\item{...}{not in use}
}
\description{
Title
}
\examples{
# simulates an example time series and estimates its trend and evolutionary wavelet spectrum

spec <- wavethresh::cns(512)
spec <- wavethresh::putD(spec, level = 8, 1 + sin(seq(from = 0, to = 2 * pi, length = 512))^2)

trend <- seq(from = 0, to = 5, length = 512)

set.seed(1)

x <- TLSW.sim(trend = trend, spec = spec)

x.TLSW <- TLSW.est(x)

plot(x.TLSW)
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Modelling
time-varying first and second-order structure of time series via wavelets
and differencing. \emph{Electronic Journal of Statistics}, 6(2), 4398-4448.

McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend
locally stationary wavelet processes. \emph{Journal of Time Series
Analysis}, 43(6), 895-917.
}
