% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TLSW.R
\name{print.TLSW}
\alias{print.TLSW}
\title{Print an Object of Class \code{TLSW}}
\usage{
\method{print}{TLSW}(x, ...)
}
\arguments{
\item{x}{A \code{TLSW} object.}

\item{...}{Other arguments.}
}
\value{
None
}
\description{
Prints a \code{TLSW} object, giving summary information.
}
\examples{
# simulates an example time series and estimates its trend and evolutionary wavelet spectrum

spec <- wavethresh::cns(512)
spec <- wavethresh::putD(spec, level = 8, 1 + sin(seq(from = 0, to = 2 * pi, length = 512))^2)

trend <- seq(from = 0, to = 5, length = 512)

set.seed(1)

x <- TLSWsim(trend = trend, spec = spec)

x.TLSW <- TLSW(x)

print(x.TLSW)

}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Modelling
time-varying first and second-order structure of time series via wavelets
and differencing. \emph{Electronic Journal of Statistics}, 6(2), 4398-4448.

McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend
locally stationary wavelet processes. \emph{Journal of Time Series
Analysis}, 43(6), 895-917.
}
\seealso{
\code{\link{TLSW}}, \code{\link{summary.TLSW}}
}
