% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TLSW.R
\name{summary.TLSW}
\alias{summary.TLSW}
\title{Summary of output provided by the \code{TLSW.est} function}
\usage{
\method{summary}{TLSW}(object, ...)
}
\arguments{
\item{object}{a \code{TLSW} object}

\item{...}{not in use}
}
\description{
Summary method for objects of class \code{TLSW}
}
