% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLSW.est.R
\name{TLSW.est}
\alias{TLSW.est}
\title{Estimate Trend and Spectrum of Trend Locally Stationary Wavelet Process}
\usage{
TLSW.est(
  x,
  do.spec.est = TRUE,
  do.trend.est = TRUE,
  gen.filter.number = 4,
  gen.family = "DaubExPhase",
  WP.filter.number = 4,
  WP.family = "DaubExPhase",
  WP.smooth = TRUE,
  WP.smooth.type = c("mean", "median", "epanechnikov")[1],
  WP.binwidth = floor(2 * sqrt(length(x))),
  WP.max.scale = floor(log2(length(x)) * 0.7),
  WP.boundary.handle = TRUE,
  WP.inv.mat = NULL,
  WP.do.diff = FALSE,
  WP.lag = 1,
  WP.diff.number = 1,
  T.est.type = c("linear", "nonlinear")[1],
  T.filter.number = 4,
  T.family = "DaubLeAsymm",
  T.transform = c("dec", "nondec")[1],
  T.boundary.handle = TRUE,
  T.max.scale = floor(log2(length(x)) * 0.7),
  T.confint = FALSE,
  T.sig.lvl = 0.05,
  T.lacf.max.lag = floor(10 * (log10(length(x)))),
  T.reps = 199,
  T.thresh.type = c("soft", "hard")[1],
  T.thresh.normal = TRUE
)
}
\arguments{
\item{x}{The time series you wish to analyse.}

\item{do.spec.est}{Logical variable, indicating whether spectral estimation is to be performed on the time series.}

\item{do.trend.est}{Logical variable, indicating whether trend estimation is to be performed on the time series.}

\item{gen.filter.number}{The index number for the wavelet that generates the
stochastic component of the time series. For the "DaubExPhase" family, the filter number can be between
1 to 10. For the "DaubLeAsymm" family, the filter number can be between 4 to 10.}

\item{gen.family}{The family of the generating wavelet. It is recommended to
use either the Daubechies Extremal Phase family, or the Daubechies Least
Asymmetric family, corresponding to the "DaubExPhase" or the "DaubLeAsymm"
options.}

\item{WP.filter.number}{The index number for the wavelet used for spectrum estimation.}

\item{WP.family}{The family of the wavelet used for spectrum estimation.}

\item{WP.smooth}{A logical variable to indicate whether smoothing is performed on the wavelet periodogram.}

\item{WP.smooth.type}{String indicating which type of smoothing to use on wavelet periodogram.
Can be \code{"mean"}, \code{"median"}, or \code{"epanechnikov"}.}

\item{WP.binwidth}{The bin width of the smoother used to smooth
the raw wavelet periodogram.}

\item{WP.max.scale}{The coarsest wavelet scale used to estimate the spectrum.
Should be a positive integer less than \eqn{J}, where \eqn{T=2^J} is the length of the
time series.}

\item{WP.boundary.handle}{Logical variable, if TRUE, the time series is
boundary corrected, to get a more accurate spectrum estimate at the
boundaries of the times series. If FALSE, no boundary correction is applied.
Recommended to use TRUE.}

\item{WP.inv.mat}{The user can pre-calculate and supply the appropriate
correction matrix used to correct the raw wavelet periodogram. If left blank,
then the correction matrix is calculated when performing spectral estimation.}

\item{WP.do.diff}{Logical variable, indicating if the time series is to be
differenced before spectral estimation is performed.}

\item{WP.lag}{The lag of differencing used, only applicable if \code{WP.do.dif = TRUE}.}

\item{WP.diff.number}{The number of differencing operations performed,
only applicable if \code{WP.do.dif = TRUE}. A first difference is recommended as default.}

\item{T.est.type}{String indicating type of wavelet thrsholding used. Can be "linear", which means
that all non-boundary wavelet coefficients are set to zero, or "nonlinear", where
each wavelet coefficient is thresholded using a time-varying, noise-dependent threshold.}

\item{T.filter.number}{The index number for the wavelet used for trend estimation.}

\item{T.family}{The family of the wavelet used for trend estimation.}

\item{T.transform}{String giving the type of wavelet transform used for trend estimation.
Can be "dec", in which case a standard (decimated) wavelet transform is used, or "nondec",
in which case a nondecimated transform is used.}

\item{T.boundary.handle}{Logical variable, if \code{TRUE}, the time series is
boundary corrected when estimating the trend.}

\item{T.max.scale}{Integer variable, selects the number of scales of the wavelet transform to
apply thresholding to for trend estimation.}

\item{T.confint}{Logical variable. If \code{TRUE}, a \code{(1-T.sig.lvl)} pointwise confidence interval is
computed for the trend estimate. For \code{T.est.type = "linear"}, this is
computed using the asymptotic distribution of the trend estimator,
whilst for \code{T.est.type = "nonlinear"}, it is computed via bootstrapping.}

\item{T.sig.lvl}{Used only if \code{T.confint = TRUE}; a numeric value
(\code{0 <= T.sig.lvl <= 1}) with which a \code{(1-T.sig.lvl)} pointwise
confidence interval for the trend estimate is generated.}

\item{T.lacf.max.lag}{Used only if \code{T.est.type = "linear"} and  \code{calc.confint = TRUE};
the maximum lag of the autocovariance to compute needed for calculating the asymptotic confidence interval.}

\item{T.reps}{Used only if \code{T.est.type = "nonlinear"} and  \code{calc.confint = TRUE}; the number of bootstrap
replications used to calculate the confidence interval.}

\item{T.thresh.type}{String variable, used only if \code{T.est.type = "nonlinear"}; the type of
thresholding function used in the trend estimation. Currently only
"soft" and "hard" are available.}

\item{T.thresh.normal}{Logical variable, used only if \code{T.est.type = "nonlinear"};
if \code{TRUE}, uses a threshold assuming the data are normally
distributed. If \code{FALSE}, uses a larger threshold to reflect non-normality.}
}
\value{
An object of class \code{"TLSW"}
}
\description{
Using wavelet-based methods, estimates the trend and evolutionary
wavelet spectrum (EWS) of a nonstationary time series. For trend estimation, either
a linear or nonlinear wavelet thresholding estimator can be computed. Spectral
 estimation can be performed with or without differencing the times series first.
}
