% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLSW.sim.R
\name{TLSW.sim}
\alias{TLSW.sim}
\title{Simulate Trend Locally Stationary Wavelet Process with Specified Trend and
Distribution of Random Innovations}
\usage{
TLSW.sim(
  trend = NULL,
  spec,
  distribution = c("norm", "pois", "exp", "chisq", "t")[1],
  rate = NULL,
  df = NULL
)
}
\arguments{
\item{trend}{A numeric vector giving the deterministic trend function to simulate.}

\item{spec}{An object of class wd which contains the spectrum for simulating
an LSW process.}

\item{distribution}{The distribution of the random variables used to
simulate the process. Can be "norm", "pois", "exp", "chisq" or
"t".}

\item{rate}{The rate parameter, only used if \code{distribution = "exp"}
or \code{distribution = "pois"}}

\item{df}{The degrees of freedom, only used if \code{distribution = "chisq"}
or \code{distribution = "t"}}
}
\value{
A vector simulated from the spectral description given in the spec
description. The returned vector will exhibit the spectral characteristics
defined by spec.
\code{\link{LSWsim}}
}
\description{
Simulates a locally stationary wavelet process given a trend function,
spectrum and distribution for the random innovations. Extension of the
\code{LSWsim} function from the \code{wavethresh} package.
}
\examples{
spec <- wavethresh::cns(1024)

spec <- wavethresh::putD(spec, level = 8, seq(from = 2, to = 8, length = 1024))

trend <- sin(pi*(seq(from = 0, to = 4, length = 1024)))

x <- TLSW.sim(trend = trend, spec = spec, distribution = "exp")

plot.ts(x)
}
