% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.estCI.R
\name{trend.estCI}
\alias{trend.estCI}
\title{Calculate Confidence Intervals Of Wavelet-Based Trend Estimate}
\usage{
trend.estCI(
  trend.est,
  lacf.est,
  filter.number = 4,
  family = "DaubLeAsymm",
  alpha = 0.95
)
}
\arguments{
\item{trend.est}{The trend estimate of the time series.}

\item{lacf.est}{The estimated local autocovariance of the time series.}

\item{filter.number}{The index of the wavelet used in the estimation
procedure.}

\item{family}{The wavelet family used in the estimation procedure.}

\item{alpha}{The significance level of the confidence interval to calculate.
Default is 0.95.}
}
\value{
A list object, containing the following objects:
\item{trend.est}{The trend estimate of the input data. } \item{trend.var}{
The variance estimate of the trend estimate. } \item{lower.conf}{ The lower
pointwise confidence interval for the trend estimate. } \item{upper.conf}{
The upper pointwise confidence interval for the trend estimate. }
}
\description{
Calculates appropriate confidence intervals for the trend
estimate, based on a given trend estimate and covariance matrix of the time
series.
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend
locally stationary wavelet processes. \emph{Journal of Time Series
Analysis}, , 43(6), 895-917.
}
\seealso{
\code{\link{wav.trend.est}}, \code{\link{lacf.calc}}
}
