\name{TrendLSW-package}
\alias{TrendLSW-package}
\alias{TrendLSW}
\docType{package}
\title{
\packageTitle{TrendLSW}
}
\description{
\packageDescription{TrendLSW}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TrendLSW}
\packageIndices{TrendLSW}
}
\author{
\packageAuthor{TrendLSW}

Maintainer: \packageMaintainer{TrendLSW}
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend locally stationary wavelet processes. Journal of Time Series Analysis.

Nason, G. P., von Sachs, R., and Kroisandt, G. (2000). Wavelet processes and adaptive estimation of the evolutionary wavelet spectrum. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 62(2):271–292.
}
\seealso{

\code{\link{ewspec.trend}}, \code{\link{ewspec.diff}}, \code{\link{wav.trend.est}}, \code{\link{wav.diff.trend.est}}

}
\keyword{ package }
