% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav.diff.trend.est.R
\name{wav.diff.trend.est}
\alias{wav.diff.trend.est}
\title{Wavelet Thresholding Trend Estimation of Time Series}
\usage{
wav.diff.trend.est(
  x,
  spec.est,
  filter.number = 4,
  thresh.type = "soft",
  normal = TRUE,
  family = "DaubLeAsymm",
  max.scale = floor(0.7 * log2(length(x))),
  boundary.handle = FALSE,
  calc.confint = FALSE,
  reps = 199,
  sig.lvl = 0.05,
  ...
)
}
\arguments{
\item{x}{The time series you want to estimate the trend function of.}

\item{spec.est}{You must supply the estimate of the evolutionary wavelet
spectrum of the time series. This is the output of the \code{ewspec.diff}
function.}

\item{filter.number}{Selects the index of the wavelet used in the estimation
procedure. For Daubechies compactly supported wavelets the filter number is
the number of vanishing moments.}

\item{thresh.type}{The type of thresholding function used. Currently only
"soft" and "hard" are available.}

\item{normal}{If TRUE, uses a threshold assuming the data are normally
distributed. If FALSE, uses a larger threshold to reflect non-normality.}

\item{family}{Selects the wavelet family to use. Recommended to only use the
Daubechies compactly supported wavelets DaubExPhase and DaubLeAsymm.}

\item{max.scale}{Selects the number of scales of the wavelet transform to
apply thresholding to. Should be a value from 1 (finest) to J-1 (coarsest),
where T=2^J is the length of the time series. Recommended to use 2J/3
scales.}

\item{boundary.handle}{Logical variable, decides if boundary handling should
be applied to the time series before estimation.}

\item{calc.confint}{Logical variable. If \code{TRUE}, a bootstrapped \code{(1-sig.lvl)}
pointwise confidence interval is computed for the trend estimate.}

\item{reps}{Used only if \code{calc.confint = TRUE}; the number of bootstrap
replications used to calcualte the confidence interval.}

\item{sig.lvl}{Used only if \code{calc.confint = TRUE}; a numeric value
(\code{0 <= sig.lvl <= 1}) with which a \code{(1-sig.lvl)} pointwise
confidence interval for the trend estimate is generated.}

\item{...}{Further arguments to be passed to the \code{\link{ewspec.diff}}
call, only to be used if \code{calc.confint = TRUE}.}
}
\value{
A vector of length \code{length(x)} containing the trend estimate.
}
\description{
Computes the wavelet thresholding trend estimate for a time
series that may be second-order nonstationary. The function calculates the
wavelet transform of the time series, thresholds the coefficients based on
an estimate of their variance, and inverts to give the trend estimate.
}
\details{
Estimates the trend function of a locally stationary time series, by
incorporating the evolutionary wavelet spectrum estimate in a wavelet
thresholding procedure. To use this function, first compute the spectral
estimate of the time series, using the function ewspec.diff.

The function works as follows:

1. The wavelet transform of the time series is calculated.

2. The wavelet coefficients are individually thresholded using the universal
threshold \eqn{\hat{\sigma}\sqrt(2 log T)}, where \eqn{\hat{sigma}^2} is an estimate of their variance. The variance
estimate is calculated using the spectral estimate, supplied by the user in
the \code{spec} argument.

3. The inverse wavelet transform is applied to obtain the final estimate.
}
\examples{
spec <- wavethresh::cns(1024, filter.number = 4)
spec <- wavethresh::putD(spec, level = 8, 1 + sin(seq(from = 0, to = 2 * pi, length = 1024))^2)

set.seed(120)

noise <- wavethresh::LSWsim(spec)
sine_trend <- -2 * sin(seq(from = 0, to = 2 * pi, length = 1024)) -
  1.5 * cos(seq(from = 0, to = pi, length = 1024))

x <- sine_trend + noise

spec.est <- ewspec.diff(x = x, family = "DaubExPhase", filter.number = 4, max.scale = 7)

trend.est <- wav.diff.trend.est(x = x, spec = spec.est)

plot.ts(x, lty = 1, col = 8)
lines(sine_trend, col = 2, lwd = 2)
lines(trend.est, col = 4, lwd = 2, lty = 2)
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Modelling
time-varying first and second-order structure of time series via wavelets
and differencing. \emph{Electronic Journal of Statistics}, 6(2), 4398-4448.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
\code{\link{ewspec.diff}}, \code{\link{wav.trend.est}}
}
