\name{z.acc}
\docType{data}
\alias{z.acc}
\title{
  Z-Axis Acceleration for Human Activity Monitoring
}
\description{

  This dataset is a section of data from Experiment 3, User 2, based on accelerometer readings from a smartphone (Reyes-Ortiz, Oneto, Sama, Parra, and Anguita (2016)), obtained from the UCI data repository (Kelly, Longjohn, and Nottingham (2024)). The data gives the time series of the acceleration along the Z-axis of an experiment participant as they perform the activities of walking up and downstairs several times.

}
\usage{
  z.acc
}
\format{
  A vector of length 6000.
}
\source{Kelly M., Longjohn R., and Nottingham, K. (2024). The UCI Machine Learning Repository. \doi{10.24432/C54G7M}.
}
\references{
Reyes-Ortiz, J. L., Oneto, L., Sama, A., Parra, X., and Anguita, D. (2016). Transition-Aware Human Activity Recognition Using Smartphones. \emph{Neurocomputing}, 171, 754–767.
}

\keyword{datasets}
\seealso{
\code{\link{z.labels}}
}
