% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atau.mat.calc.R
\name{Atau.mat.calc}
\alias{Atau.mat.calc}
\title{Lagged Autocorrelation Wavelet Inner Product Calculation}
\usage{
Atau.mat.calc(J, filter.number = 1, family = "DaubExPhase", lag = 1)
}
\arguments{
\item{J}{The dimension of the matrix required. Should be a positive integer.}

\item{filter.number}{The index of the wavelet used to compute the inner
product matrix.}

\item{family}{The family of wavelet used to compute the inner product
matrix.}

\item{lag}{The lag of matrix to calculate. A lag of 0 corresponds to the
standard A matrix.}
}
\value{
A J-dimensional square matrix giving the lagged inner product
autocorrelation wavelet matrix.
}
\description{
Computes the matrix of lagged autocorrelation wavelet inner
products.
}
\details{
Computes the lagged inner product matrix of the discrete
non-decimated autocorrelation wavelets. This matrix is used in the
calculation to correct the wavelet periodogram of the differenced time
series. With \code{lag} \eqn{= \tau}, the matrix returned is the one called \eqn{A^\tau} in McGonigle et al. (2022).
}
\examples{
A1 <- Atau.mat.calc(J = 5, filter.number = 1, family = "DaubExPhase", lag = 1)
}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Modelling
time-varying first and second-order structure of time series via wavelets
and differencing. \emph{Electronic Journal of Statistics}, 6(2), 4398-4448.
}
\seealso{
\link{TLSW}
}
