% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TLSW.R
\name{plot.TLSW}
\alias{plot.TLSW}
\title{Plot Trend and/or Spectrum Information in a \code{TLSW} Object}
\usage{
\method{plot}{TLSW}(x, plot.type = c("trend", "spec"), trend.plot.args, spec.plot.args, ...)
}
\arguments{
\item{x}{A \code{TLSW} object}

\item{plot.type}{A string object indicating what is to be plotted. Can be:
\itemize{
\item{\code{"trend"}}{: will plot the trend estimate only.}
\item{\code{"spec"}}{: will plot the spectral estimate only.}
\item{\code{c("trend", "spec")}}{: the default value will plot both the trend and spectral estimate.}
}}

\item{trend.plot.args}{A list object, that includes any choices for the graphical parameters used for plotting the trend estimate.}

\item{spec.plot.args}{A list object, that includes any choices for the graphical parameters used for plotting the spectral estimate.}

\item{...}{Any additional arguments that will be applied to the graphical parameters of both the trend and spectrum plotting.}
}
\description{
Plots information contained within a \code{TLSW} object.
}
\examples{
# Simulates an example time series and estimates its trend and evolutionary wavelet spectrum.
# Then plots both estimates.

spec <- matrix(0, nrow = 9, ncol = 512)

spec[1,] <- 1 + 2*sin(seq(from = 0, to = 2 * pi, length = 512))^2

trend <- seq(from = 0, to = 5, length = 512) + sin(seq(from = 0, to = 2 * pi, length = 512))

set.seed(1)

x <- TLSWsim(trend = trend, spec = spec)

x.TLSW <- TLSW(x)

plot(x.TLSW, trend.plot.args = list(ylab = "Simulated Data"))

}
\references{
McGonigle, E. T., Killick, R., and Nunes, M. (2022). Modelling
time-varying first and second-order structure of time series via wavelets
and differencing. \emph{Electronic Journal of Statistics}, 6(2), 4398-4448.

McGonigle, E. T., Killick, R., and Nunes, M. (2022). Trend
locally stationary wavelet processes. \emph{Journal of Time Series
Analysis}, 43(6), 895-917.
}
